requirements.txt
# Dev environment
pip
autopep8
python-dotenv

# App
google-generativeai # Client library for the Gemini API
Pillow # PIL image-loading library
flask

# Security
flask_cors
google-cloud-secret-manager